+++ CHANGELOG.md
@@ -1,3 +1,17 @@
-## [0.0.1] - TODO: Add release date.
+## [0.1.0] - 2/7/2020.
 
-* TODO: Describe initial release.
+* Refactor code.
+* Add point position.
+* Remove margin bottom (instead that using insets)
+* Remove PointTabBar Widget
+
+* Update change log.
+* Update description.
+
+## [0.0.2] - 2/7/2020.
+
+* Refactor code.
+
+## [0.0.1] - 2/7/2020.
+
+* Release.